Re-usable ContainerTestCase class
