+++ snapcraft/tests/commands/test_snap.py
@@ -31,10 +31,9 @@ from testtools.matchers import (
 
 from snapcraft.main import main
 from snapcraft import tests
-from snapcraft.tests.test_lxd import check_output_side_effect
 
 
-class SnapCommandContainerized(tests.TestCase):
+class SnapCommandContainerized(tests.ContainerTestCase):
 
     yaml_template = """name: snap-test
 version: 1.0
@@ -49,21 +48,6 @@ parts:
     part1:
       plugin: nil
 """
-
-    def setUp(self):
-        super().setUp()
-        patcher = mock.patch('snapcraft.internal.lxd.check_call')
-        self.check_call_mock = patcher.start()
-        self.addCleanup(patcher.stop)
-
-        patcher = mock.patch('snapcraft.internal.lxd.check_output')
-        self.check_output_mock = patcher.start()
-        self.check_output_mock.side_effect = check_output_side_effect()
-        self.addCleanup(patcher.stop)
-
-        patcher = mock.patch('snapcraft.internal.lxd.sleep', lambda _: None)
-        patcher.start()
-        self.addCleanup(patcher.stop)
 
     def make_snapcraft_yaml(self, n=1):
         super().make_snapcraft_yaml(self.yaml_template)
